requirements.txt
aioredis==1.3.1
altgraph==0.17
apt-xapian-index==0.49
asgiref==3.4.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
autopep8==1.5.7
beautifulsoup4==4.9.3
blinker==1.4
cbor==1.0.0
certifi==2019.11.28
channels==3.0.4
channels-redis==3.3.0
chardet==3.0.4
click==8.0.3
colorama==0.4.3
command-not-found==0.3
configparser==5.0.2
constantly==15.1.0
crayons==0.4.0
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
Cython==0.29.14
daphne==3.0.2
dbus-python==1.2.16
defusedxml==0.7.1
deluge==2.0.3
diff-match-patch==20200713
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.8
django-debug-toolbar==3.2.2
django-debug-toolbar-request-history==0.1.4
django-filter==2.4.0
django-import-export==2.6.1
djangorestframework==3.12.4
entrypoints==0.3
et-xmlfile==1.1.0
filetype==1.0.8
GeoIP==1.3.2
gevent==21.8.0
gpg===1.13.1-unknown
greenlet==1.1.1
gunicorn==20.1.0
h2==3.2.0
hiredis==2.0.0
hpack==3.0.0
httplib2==0.14.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
imutils==0.5.4
incremental==16.10.1
ipython-genutils==0.2.0
itemadapter==0.3.0
itemloaders==1.0.4
jmespath==0.10.0
jsonschema==3.2.0
jupyter-core==4.7.1
keyring==18.0.1
labelImg==1.8.6
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libcst==0.3.23
lxml==4.6.3
lz4==3.0.2+dfsg
Mako==1.1.0
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==1.1.0
msgpack==1.0.2
mypy-extensions==0.4.3
mysqlclient==2.0.3
nbformat==5.1.3
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
opencv-python==4.5.4.58
openpyxl==3.0.9
parsel==1.6.0
pathspec==0.9.0
pbr==5.4.5
pexpect==4.6.0
pigar==0.10.0
Pillow==7.0.0
priority==1.3.0
Protego==0.1.16
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycln==1.1.0
pycodestyle==2.7.0
pycrypto==2.6.1
pycups==1.9.73
PyDispatcher==2.0.5
pygame==1.9.6
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyrsistent==0.18.0
pytesseract==0.3.8
python-apt==2.0.0+ubuntu0.20.4.6
python-debian===0.1.36ubuntu1
python-decouple==3.5
python-libtorrent==1.1.13
python-snap7==1.1
python-snappy==0.5.3
PyTrie==0.2
pytz==2021.1
pyxdg==0.26
PyYAML==5.3.1
qrcode==7.3.1
queuelib==1.6.1
rencode==1.0.6
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
scipy==1.7.3
Scrapy==2.5.0
SecretStorage==2.3.1
selenium==3.141.0
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.16.0
sip==4.19.21
six==1.14.0
soupsieve==2.2.1
sqlparse==0.4.1
supervisor==4.1.0
systemd-python==234
tablib==3.0.0
toml==0.10.2
traitlets==5.1.0
Twisted==18.9.0
txaio==2.10.0
typer==0.4.0
typing-extensions==4.0.0
typing-inspect==0.7.1
u-msgpack-python==2.1
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unipath==1.1
urllib3==1.25.8
usb-creator==0.3.7
w3lib==1.22.0
wadllib==1.3.3
webdriver-manager==3.4.2
websocket==0.2.1
websocket-client==1.2.1
websockets==9.1
whitenoise==5.3.0
wsaccel==0.6.2
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zope.event==4.5.0
zope.interface==4.7.1
